refactoring streamlit app
